add jagged rock + revert old changes to tiles, objects and decals.
recent commit was done to supertux that added jagged rock tileset. this commit adds that commit here but unfortunately, due to the fact that i'm too lazy, i basically added back most of the images that i deleted in the past. :( this means the file size is gonna be larger now which sucks i'd have to do this at some point considering how big the jagged rock stuff is since it adds even more stuff than just jagged rock which makes it more important to add than if it was only jagged rock.